# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "高考志愿AI顾问",
    "description": "AI 志愿规划服务，支持高考、考研和职业方向分析",
    "version": "3.1.0"
  },
  "paths": {
...
    "/api/v1/data/schools": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Get Schools",
        "description": "Search schools by name, province, or level with cursor-based pagination.",
        "operationId": "get_schools_api_v1_data_schools_get",
        "parameters": [
          {
            "name": "school_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "学校名称关键词",
              "title": "School Name"
            },
            "description": "学校名称关键词"
          },
          {
            "name": "province",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "省份",
              "title": "Province"
            },
            "description": "省份"
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "学校层次",
              "title": "Level"
            },
            "description": "学校层次"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "每页数量",
              "default": 50,
              "title": "Limit"
            },
            "description": "每页数量"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "游标（上一页的 next_cursor）",
              "title": "Cursor"
            },
            "description": "游标（上一页的 next_cursor）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}